requirements.txt
requests
dnspython
python-whois
httpx
colorama
python-nmap
